refactor: moved store
