--- cognite/neat/graph/__init__.py
@@ -1,3 +1,0 @@
-from .stores import NeatGraphStore
-
-__all__ = ["NeatGraphStore"]
