+++ cognite/neat/stores/_base.py
