+++ tests/tests_unit/graph/test_graph_read.py
@@ -2,8 +2,8 @@ from cognite.client.data_classes import AssetList
 from cognite.client.testing import monkeypatch_cognite_client
 
 from cognite.neat.graph.extractors import AssetsExtractor
-from cognite.neat.graph.stores import NeatGraphStore
 from cognite.neat.rules.importers import InferenceImporter
+from cognite.neat.stores import NeatGraphStore
 from tests.config import CLASSIC_CDF_EXTRACTOR_DATA
