+++ tests/tests_unit/test_graph/test_loaders/test_rdf2dms.py
@@ -1,8 +1,8 @@
 import pytest
 
 from cognite.neat.graph.loaders import DMSLoader
-from cognite.neat.graph.stores import NeatGraphStore
 from cognite.neat.rules.importers import InferenceImporter
+from cognite.neat.stores import NeatGraphStore
 from tests.data import car
